+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Changes to Calva Backseat Driver
 ## [Unreleased]
 
 - [Add human intelligence tool, for the agent to ask the human for guidance](https://github.com/BetterThanTomorrow/calva-backseat-driver/issues/17)
+- [Place MCP stdio wrapper script in user home Calva config directory](https://github.com/BetterThanTomorrow/calva-backseat-driver/issues/15)
 
 ## [v0.0.13] - 2025-05-28